Rename builder module to batch
